Add some strange things
Я тут просто типа порядок навел, избавился от всех ошибок перед тем как
наделать новых, похорошему надо было это в ветке develop делать но мне
насрать.

Card: S01E11
