+++ .flowconfig
@@ -1,13 +1,19 @@
 [include]
 .*/src/.*
+.*/node_modules/.*
 
 [ignore]
-.*/node_modules/.*
 .*/implements/.*
 .*/public/.*
 .*/styleguide/.*
 
+# Ignore some libs
+.*/node_modules/stylelint/.*
+.*/node_modules/findup/.*
+
+
 [options]
+module.system.node.resolve_dirname=node_modules
 module.file_ext=.js
 module.file_ext=.jsx
 module.file_ext=.json
