+++ package.json
@@ -15,7 +15,7 @@
     "build": "better-npm-run build",
     "lint": "npm run eslint & npm run stylelint",
     "stylelint": "stylelint src/**/*.css",
-    "eslint": "eslint src",
+    "eslint": "eslint --ext .jsx --ext .js src",
     "flow": "flow --color always",
     "styleguide": "better-npm-run styleguide",
     "styleguide:build": "better-npm-run styleguide:build"
@@ -102,7 +102,7 @@
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-json": "^1.2.0",
     "eslint-plugin-jsx-a11y": "^2.2.3",
-    "eslint-plugin-react": "^6.4.1",
+    "eslint-plugin-react": "^6.10.3",
     "extract-text-webpack-plugin": "2.0.0-beta",
     "favicons-webpack-plugin": "^0.0.7",
     "file-loader": "^0.9.0",
