+++ src/tools/styleguide/Playground/Playground.css
@@ -56,7 +56,12 @@
   transition: width 0.5s, height 0.5s;
 
   &_isResize {
-    transition: none;
+    /*
+      NOTE: Remove when i fix this shit;
+      https://github.com/konstantin24121/stylelint-performance-animation/issues/2
+    */
+    /* stylelint-disable-next-line plugin/no-low-performance-animation */
+    transition: inherit;
   }
 }
