+++ src/tools/styleguide/StyleGuide/index.js
@@ -1,1 +1,1 @@
-export { default } from './StyleGuide.js';
+export { default } from './StyleGuide.jsx';
