+++ src/tools/styleguide/Toolbar/Toolbar.css
@@ -7,7 +7,6 @@
   justify-content: space-between;
   padding: 0 1rem;
 
-
   @media (min-width: 768px) {
     width: 50px;
     flex-direction: column;
